+++ scripts/deleteFile.php
@@ -23,9 +23,38 @@ if (time() - $_SESSION['CREATED'] > 1800) {
     $_SESSION['CREATED'] = time();  // update creation time
 }
 
-$file = "../uploads/".$_SESSION['login_user']."/".$_POST['filename'];
-unlink($file);
-echo "The file has been deleted.";
-header("Location: ../home.php");
-die();
+if(!function_exists('hash_equals')) {
+  function hash_equals($str1, $str2) {
+    if(strlen($str1) != strlen($str2)) {
+      return false;
+    } else {
+      $res = $str1 ^ $str2;
+      $ret = 0;
+      for($i = strlen($res) - 1; $i >= 0; $i--) $ret |= ord($res[$i]);
+      return !$ret;
+    }
+  }
+}
+
+if (!empty($_POST['CSRFToken'])) {
+    if (hash_equals($_SESSION['token'], $_POST['CSRFToken'])) {
+         // Proceed to process the form data
+
+		$file = "../uploads/".$_SESSION['login_user']."/".$_POST['filename'];
+		unlink($file);
+		echo "The file has been deleted.";
+		header("Location: ../home.php");
+		die();
+
+	} else {
+         // Log this as a warning and keep an eye on these attempts
+        echo '<script type="text/javascript"> confirm("Invalid CSRF Token");';
+        echo 'window.location= "../home.php";</script>';
+        throw new Exception("Invalid CSRF token.");
+    }
+}
+else{
+    echo '<script type="text/javascript">';
+    echo 'window.location= "../home.php";</script>';
+}
 ?>
