+++ scripts/upload.php
@@ -23,35 +23,64 @@ if (time() - $_SESSION['CREATED'] > 1800) {
     $_SESSION['CREATED'] = time();  // update creation time
 }
 
-$target_dir = "../uploads/".$_SESSION['login_user']."/";
-
-$target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
-$uploadOk = 1;
-$fileType = pathinfo($target_file,PATHINFO_EXTENSION);
-
-// Check if file already exists
-if (file_exists($target_file)) {
-    echo '<script type="text/javascript"> confirm("File with same name already exists");</script>';
-    $uploadOk = 0;
-}
-// Check file size
-//if ($_FILES["fileToUpload"]["size"] > 500000) {
-//    echo "Sorry, your file is too large. (>500kb)";
-//    $uploadOk = 0;
-//}
-
-// Check if $uploadOk is set to 0 by an error
-if ($uploadOk == 0) {
-    echo '<script type="text/javascript"> confirm("Sorry, your file was not uploaded");';
-    echo 'window.location= "../home.php";</script>';
-// if everything is ok, try to upload file
-} else {
-    if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
-        header("Location: ../home.php");
-        die();
+if(!function_exists('hash_equals')) {
+  function hash_equals($str1, $str2) {
+    if(strlen($str1) != strlen($str2)) {
+      return false;
     } else {
-        echo '<script type="text/javascript"> confirm("Sorry, there was an error uploading your file");';
-        echo 'window.location= "../home.php";</script>';
+      $res = $str1 ^ $str2;
+      $ret = 0;
+      for($i = strlen($res) - 1; $i >= 0; $i--) $ret |= ord($res[$i]);
+      return !$ret;
     }
+  }
+}
+
+if (!empty($_POST['CSRFToken'])) {
+    if (hash_equals($_SESSION['token'], $_POST['CSRFToken'])) {
+         // Proceed to process the form data
+
+        $target_dir = "../uploads/".$_SESSION['login_user']."/";
+
+        $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
+        $uploadOk = 1;
+        $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
+
+        // Check if file already exists
+        if (file_exists($target_file)) {
+            echo '<script type="text/javascript"> confirm("File with same name already exists");</script>';
+            $uploadOk = 0;
+        }
+        // Check file size
+        //if ($_FILES["fileToUpload"]["size"] > 500000) {
+        //    echo "Sorry, your file is too large. (>500kb)";
+        //    $uploadOk = 0;
+        //}
+
+        // Check if $uploadOk is set to 0 by an error
+        if ($uploadOk == 0) {
+            echo '<script type="text/javascript"> confirm("Sorry, your file was not uploaded");';
+            echo 'window.location= "../home.php";</script>';
+        // if everything is ok, try to upload file
+        } else {
+            if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
+                header("Location: ../home.php");
+                die();
+            } else {
+                echo '<script type="text/javascript"> confirm("Sorry, there was an error uploading your file");';
+                echo 'window.location= "../home.php";</script>';
+            }
+        }
+
+    } else {
+         // Log this as a warning and keep an eye on these attempts
+        echo '<script type="text/javascript"> confirm("Invalid CSRF Token");';
+        echo 'window.location= "../home.php";</script>';
+        throw new Exception("Invalid CSRF token.");
+    }
+}
+else{
+    echo '<script type="text/javascript">';
+    echo 'window.location= "../home.php";</script>';
 }
 ?>
